fix(WelcomeCard): fix image loading in draw function
The background image was not being loaded correctly in the draw function of WelcomeCard.js, which was causing the image to not be displayed properly. This commit fixes this issue by correctly loading the image in the draw function.

Changes made:

Moved image loading to the draw function
Fixed context.closePath() in draw function
File(s) modified:
WelcomeCard.js